[BE] feat: 팔로우기반, 마이페이지기반 에서 태그검색

+++ backend/src/main/java/a402/FaST/model/dto/TagRequestDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 public class TagRequestDto {
 
-   private int tagId;
+   private Long tagId;
    private int userId;
    private List<String> tags;
